sta-app/requirements-ui.txt
websockets==10.4
requests==2.28.2
pulsar-client==2.10.2
pulsar-client[avro]==2.10.2